Finished first version of the java 'executor'
There is a good chance there will be a load of bugs in here,
since the executor can't be fully tested yet. We need db
order tables before testing can really begin.

A simple order class file was also created for the executor. It may
need to be expanded by whoever writes the buy and sell orders.